refactor: Move Parser combinators to that mod
